+++ python/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2025-04-21
+### Added
+- Added support for PayPal Invoices.
+- Added LangChain integration with `langchain.PayPalToolkit.get_tools()` for LangChain agents.
+
 ## [1.0.1] - 2025-04-18
-### Bug Fixes
+### Fixed
 - Handled `204 No Content` response formatting issue.
 - Fixed parameters formatting issue.